Set up an om xml document for oai-pmh oai_dc docs
om (opinionated metadata) is a gem for translating between
XML schemas and Ruby objects (it is used extensively by Hydra, which is
also a Blacklight application)

Instead of accessing the om terminology for the Record#content
indirectly, we use the Rails 3.x .serialize. By adding a special
Serializer class, we can access the terminology (and modify, save, etc)
directly on Record#content.