Split OAI-PMH record metadata from the payload
We want to think about supporting generic OAI-PMH provided content. The
structure of an OAI-PMH document is defined in the specification, but
the payload of <metadata> is provider specific.

Here, we create the model OaiPmhRecord, which can be mixed into any
class that provides the appropriate terminology helpers:

.set_oai_metadata_terminology : add metadata-specific terminology

.add_oai_metadata_terminology_proxies : add top-level terminology
accessors to the OAI document